backend/pyproject.toml
[tool.poetry]
name = "countonme-backend"
version = "0.1.0"
description = "CountOnMe backend (FastAPI + async SQLAlchemy + Postgres)"
authors = ["CountOnMe"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
sqlalchemy = "^2.0.0"
asyncpg = "^0.30.0"
psycopg = { extras = ["binary"], version = "^3.2.0" }
alembic = "^1.13.0"
pydantic-settings = "^2.7.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.0"
pytest = "^8.3.0"
pytest-asyncio = "^0.25.0"
httpx = "^0.28.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

